Changed 4 files:
game_rules_class.py
graphic_die_class.py
graphic_die_game.py
graphic_die_interface.py

I have added 2 more buttons to the game that allow the player to start a new game at any time and a button that allows them to quit without using the esc key or close the window.

I still need to find a way to check for a lose state. I have implemented a couple of helper functions, but this is still not working.